requirements.txt
Django
djangorestframework
django-allauth
dj-rest-auth
markdown       
django-filter
psycopg2-binary
cx_Oracle
dadata
mammoth